+++ fetch.js
@@ -21,24 +21,28 @@ class FetchSession extends Session {
 		super( apiUrl, defaultParams, defaultOptions );
 	}
 
-	async internalGet( params, userAgent ) {
+	async internalGet( params, headers ) {
 		const url = new URL( this.apiUrl );
 		url.search = new URLSearchParams( params );
+		const { 'user-agent': userAgent, ...otherHeaders } = headers;
 		const response = await fetch( url, {
 			headers: {
+				...otherHeaders,
 				'api-user-agent': userAgent,
 			},
 		} );
 		return transformResponse( response );
 	}
 
-	async internalPost( urlParams, bodyParams, userAgent ) {
+	async internalPost( urlParams, bodyParams, headers ) {
 		const url = new URL( this.apiUrl );
 		url.search = new URLSearchParams( urlParams );
+		const { 'user-agent': userAgent, ...otherHeaders } = headers;
 		const response = await fetch( url, {
 			method: 'POST',
 			body: new URLSearchParams( bodyParams ),
 			headers: {
+				...otherHeaders,
 				'api-user-agent': userAgent,
 			},
 		} );